Make the Steam and SteamVR paths configurable
Add steamPath and steamVrPath to config.bat, reference them in mixedvr-manager.bat.

Possibly depending on when you first installed it, Steam seems to call the SteamVR directory "SteamVR" or "OpenVR". It's also possible for this directory to be in a different root to the "config" folder, if you use Steam's alternate drive functionality.